requirements.txt
click==7.1.2
colorzero==1.1
inputs==0.5
itsdangerous==1.1.0
MarkupSafe==1.1.1
numpy
opentrons==4.4.0
opencv-python
pandas
Pillow==9.3.0
protobuf==3.12.2
pygame
pyserial==3.5
python-dateutil==2.8.1
pytz==2020.4
requests==2.28.1
Rx==3.1.1
six==1.15.0
tkhtmlview==0.1.1.post5
Werkzeug==1.0.1
xlrd==2.0.1
zaber-motion==1.5.1
zaber-motion-bindings-linux==4.7.0
zaber-motion-bindings-windows==1.5.1